+++ enc/utf_32be.c
@@ -155,7 +155,7 @@ static UChar*
 utf32be_left_adjust_char_head(const UChar* start, const UChar* s, const UChar* end,
 			      OnigEncoding enc ARG_UNUSED)
 {
-  int rem;
+  ptrdiff_t rem;
 
   if (s <= start) return (UChar* )s;
